+++ src/main/java/com/aleksander/restaurant/reservation/dto/stats/DailyReservationStatsDTO.java
@@ -1,0 +1,13 @@
+package com.aleksander.restaurant.reservation.dto.stats;
+
+import java.time.LocalDate;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@Data
+@AllArgsConstructor
+public class DailyReservationStatsDTO {
+    private LocalDate date;
+    private long reservationsCount;
+}