+++ src/main/java/com/aleksander/restaurant/reservation/service/StatsService.java
@@ -2,9 +2,14 @@ package com.aleksander.restaurant.reservation.service;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Map;
+import java.util.stream.Collectors;
 
 import org.springframework.stereotype.Service;
 
+import com.aleksander.restaurant.reservation.dto.stats.DailyReservationStatsDTO;
 import com.aleksander.restaurant.reservation.dto.stats.ReservationCountDTO;
 import com.aleksander.restaurant.reservation.dto.stats.RestaurantStatsDTO;
 import com.aleksander.restaurant.reservation.model.ReservationStatus;
@@ -51,5 +56,27 @@ public class StatsService {
                 .days(days)
                 .reservationsCount(count)
                 .build();
+    }
+
+    public List<DailyReservationStatsDTO> getDailyReservationsCountLastNDaysStats(int days) {
+        LocalDate today = LocalDate.now();
+        LocalDate startDate = today.minusDays(days - 1);
+
+        List<Object[]> counts = reservationRepository.countReservationsGroupedByDate(startDate.atStartOfDay(),
+                today.plusDays(1).atStartOfDay());
+
+        Map<LocalDate, Long> countsMap = counts.stream()
+                .collect(Collectors.toMap(
+                        obj -> ((LocalDateTime) obj[0]).toLocalDate(),
+                        obj -> (Long) obj[1]));
+
+        List<DailyReservationStatsDTO> result = new ArrayList<>();
+        for (int i = 0; i < days; i++) {
+            LocalDate day = startDate.plusDays(i);
+            long count = countsMap.getOrDefault(day, 0L);
+            result.add(new DailyReservationStatsDTO(day, count));
+        }
+
+        return result;
     }
 }